Moved EmailSubscriptionCheckbox to Authenticators
